requirements.txt
fastapi
fastapi-mcp
google-generativeai
langchain
langchain-community
langchain-experimental
loguru
numpy
pydantic
pydantic-settings
psycopg2-binary
python-dotenv
qdrant-client~=1.8.0
sentence-transformers
transformers
uvicorn[standard]
volcengine-python-sdk
accelerate